refactor(cosmos): make TxRaw assembly sign-mode-agnostic
- makeTxRaw takes an explicit signMode: SignMode param so AMINO and the
  upcoming DIRECT pipeline share the same TxRaw builder. No default value —
  a wrong mode on the AuthInfo side silently invalidates the signature, so
  the choice is forced at every call site.
- Extract resolveAccount + resolvePublicKey from sign-cosmos-amino into a
  new signer-helpers module; DIRECT reuses the same account/pubkey logic
- Standardize the Ledger guard message to mention "(Phase 7)" so both
  signer-helpers and LedgerKeyring.signRaw agree on the forward-looking label
- AMINO GOLDEN unchanged — bit-for-bit identical txBytes
